+++ CHANGELOG.md
@@ -5,6 +5,8 @@
  - Fix issue determining PID setpoint unit on LCDs
  - Fix issue displaying IP address on LCD
  - Fix issue with client activating controllers ([#532](https://github.com/kizniche/mycodo/issues/532))
+ - Fix issue with Linux Command Input ([#537](https://github.com/kizniche/mycodo/issues/537))
+ - Fix German "Output" translation
 
 
 ## 6.3.9 (2018-09-18)