+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 6.3.10 (Unreleased)
+## 6.4.0 (Unreleased)
 
 ### Features
 
@@ -11,6 +11,7 @@
  - Fix issue with client activating controllers ([#532](https://github.com/kizniche/mycodo/issues/532))
  - Fix issue with Linux Command Input ([#537](https://github.com/kizniche/mycodo/issues/537))
  - Fix issue with installing internal dependencies (e.g. pigpiod) ([#538](https://github.com/kizniche/mycodo/issues/538))
+ - Potential fix for Miflora input ([#540](https://github.com/kizniche/mycodo/issues/540))
  - Fix German "Output" translation